components/kubeflow/pytorch-launcher/requirements.txt
pyyaml
kubernetes
kubeflow-pytorchjob
